feat: Add welcome screen and release v1.0.1
Added a welcome screen with a logo animation for the first launch of the application. The user is prompted to select a music folder, after which the main application window opens. This improves the initial user experience. Also, the application version has been updated to 1.0.1 and a release build has been created.